# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /officials:
    get:
      operationId: listOfficials
      tags: [officials]
      summary: List public officials and staff
      parameters:
        - name: role
          in: query
          required: false
          description: sindaco | assessore | consigliere | dirigente | dipendente
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of officials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Official"
    post:
      operationId: createOfficial
      tags: [officials]
      summary: Create a public official with optional profile data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficialInput"
      responses:
        "201":
          description: Created official profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /officials/{id}:
    get:
      operationId: getOfficial
      tags: [officials]
      summary: Get a single official's full transparency profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Official profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: updateOfficial
      tags: [officials]
      summary: Update an official and replace their profile data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficialInput"
      responses:
        "200":
          description: Updated official profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficialProfile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /delibere/{id}/votes:
    get:
      operationId: getDeliberaVotes
      tags: [officials]
      summary: Voting breakdown for a single delibera
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Delibera voting breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliberaVotes"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MacrotemaKey:
      type: string
      description: Ambito di spesa (macrotema) di un contratto
      enum:
        - ambiente
        - scuole
        - strade
        - sociale
        - cultura
        - mobilita
        - altro
...
    Publication:
      type: object
      required:
        - id
        - publicId
        - progressivo
        - tipologia
        - category
        - oggetto
        - cups
        - isPnrr
        - isNew
        - firstSeenAt
        - macrotema
        - presentation
        - publicSafety
      properties:
        id:
          type: integer
        publicId:
          type: string
          description: Stable cross-surface identifier derived from progressivo.
        progressivo:
          type: string
        tipologia:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        provenienza:
          type: ["string", "null"]
        oggetto:
          type: string
        dataAtto:
          type: ["string", "null"]
        pubStart:
          type: ["string", "null"]
        pubEnd:
          type: ["string", "null"]
        numRegSet:
          type: ["string", "null"]
        numRegGen:
          type: ["string", "null"]
        cups:
          type: array
          items:
            type: string
        pnrrMission:
          type: ["string", "null"]
        isPnrr:
          type: boolean
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/PublicationAttachment"
        isNew:
          type: boolean
        firstSeenAt:
          type: string
        macrotema:
          $ref: "#/components/schemas/MacrotemaKey"
          description: Spending area automatically classified from the act text
        brief:
          type: ["string", "null"]
          description: AI-generated plain-language summary ("In breve")
        briefManual:
          type: boolean
          description: >-
            True when the "In breve" summary was written/curated by hand; the
            automatic batch never overwrites it.
        briefGeneratedAt:
          type: ["string", "null"]
          description: >-
            When the "In breve" summary was last generated or edited (ISO 8601).
        odgMacrotemi:
          type: array
          description: Per convocazioni — macrotemi aggregati dai punti ODG (può coprire più temi). Vuoto per gli altri tipi di atti.
          items:
            type: string
        presentation:
          $ref: "#/components/schemas/PublicationPresentation"
        publicSafety:
          $ref: "#/components/schemas/PublicActSafety"
...
    PublicationPresentation:
      type: object
      required:
        - display_title
        - action_id
        - action_label
        - search_text
        - area_theme
        - standardisation
      properties:
        display_title:
          type: string
        action_id:
          type: ["string", "null"]
        action_label:
          type: ["string", "null"]
        search_text:
          type: string
        area_theme:
          $ref: "#/components/schemas/PublicationAreaTheme"
        standardisation:
          $ref: "#/components/schemas/PublicationStandardisationAudit"
    PublicationAreaThemeEvidence:
      type: object
      required: [rule_id, input_field, matched_terms]
      properties:
        rule_id:
          type: string
        input_field:
          type: string
        matched_terms:
          type: array
          items:
            type: string
...
    PublicationAreaTheme:
      type: object
      required:
        - schema_version
        - taxonomy_id
        - taxonomy_version
        - theme_id
        - theme_label
        - confidence
        - basis
        - rule_id
        - evidence
        - null_reason
        - override
      properties:
        schema_version:
          type: string
        taxonomy_id:
          type: string
        taxonomy_version:
          type: string
        theme_id:
          type: ["string", "null"]
        theme_label:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
          enum: [high, medium, null]
        basis:
          type: string
          enum: [deterministic_rule, manual_override, fallback]
        rule_id:
          type: ["string", "null"]
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/PublicationAreaThemeEvidence"
        null_reason:
          type: ["string", "null"]
          enum:
            - input_withheld_for_privacy
            - input_missing
            - not_classified
            - ambiguous_match
            - null
        override:
          $ref: "#/components/schemas/PublicationAreaThemeOverride"
...
    PublicActSafety:
      type: object
      required:
        - policy_id
        - policy_version
        - standardisation_profile_id
        - standardisation_profile_version
        - public_visibility
        - privacy_risk
        - reason
        - projection_schema_version
        - attachments_attested
        - markdown_attested
        - attestation_status
        - attestation_reason
        - attestation_source
        - attested_at
        - source_fingerprint_verified
      properties:
        policy_id:
          type: string
        policy_version:
          type: string
        standardisation_profile_id:
          type: string
        standardisation_profile_version:
          type: string
        public_visibility:
          type: string
          enum:
            - publishable
            - publishable_with_minimisation
            - metadata_only
        privacy_risk:
          type: string
          enum: [low, medium, high]
        reason:
          type: ["string", "null"]
        projection_schema_version:
          type: string
        attachments_attested:
          type: boolean
        markdown_attested:
          type: boolean
        attestation_status:
          type: string
          enum: [valid, legacy_missing, invalid, stale, source_changed]
        attestation_reason:
          type: ["string", "null"]
          enum:
            - null
            - missing
            - invalid_schema
            - invalid_timestamp
            - invalid_source
            - invalid_fingerprint
            - stale_policy
            - stale_profile
            - invalid_decision
            - invalid_presentation
            - source_changed
        attestation_source:
          type: ["string", "null"]
          enum: [albo_ingestion, null]
        attested_at:
          type: ["string", "null"]
          format: date-time
        source_fingerprint_verified:
          type: boolean
    PublicationAttachment:
      type: object
      required:
        - name
        - tipo
        - officialUrl
        - storagePath
        - contentType
        - size
      properties:
        name:
          type: string
        tipo:
          type: string
        officialUrl:
          type: string
          description: Direct download link to the specific document on the official portal
        storagePath:
          type: ["string", "null"]
          description: Path of the locally-archived copy, or null if not archived
        contentType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
...
    Official:
      type: object
      required: [id, name, slug, role, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        role:
          type: string
          description: sindaco | assessore | consigliere | dirigente | dipendente
        roleTitle:
          type: ["string", "null"]
        group:
          type: ["string", "null"]
        status:
          type: string
          description: in_carica | cessato
        appointmentDate:
          type: ["string", "null"]
        biography:
          type: ["string", "null"]
    OfficialActivity:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
    OfficialRemuneration:
      type: object
      required: [id, year, type]
      properties:
        id:
          type: integer
        year:
          type: integer
        amount:
          type: ["number", "null"]
        type:
          type: string
        note:
          type: ["string", "null"]
    OfficialDeclaration:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
    OfficialVote:
      type: object
      required: [publicationId, vote, oggetto]
      properties:
        publicationId:
          type: integer
        vote:
          type: string
          enum: [favorevole, contrario, astenuto, assente]
        oggetto:
          type: string
        numRegGen:
          type: ["string", "null"]
        subcategory:
          type: ["string", "null"]
        dataAtto:
          type: ["string", "null"]
    OfficialOrgano:
      type: object
      required:
        [
          id,
          type,
          name,
          slug,
          membershipRole,
          termLabel,
          startDate,
          endDate,
          sourceLabel,
          sourceUrl,
          notes,
          isCurrent,
        ]
      properties:
        id:
          type: integer
        type:
          type: string
          description: consiglio | giunta | commissione
        name:
          type: string
        slug:
          type: string
        membershipRole:
          type: ["string", "null"]
        termLabel:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        sourceLabel:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isCurrent:
          type: boolean
    OfficialProfile:
      allOf:
        - $ref: "#/components/schemas/Official"
        - type: object
          required: [activities, remunerations, declarations, votes, organi]
          properties:
            activities:
              type: array
              items:
                $ref: "#/components/schemas/OfficialActivity"
            remunerations:
              type: array
              items:
                $ref: "#/components/schemas/OfficialRemuneration"
            declarations:
              type: array
              items:
                $ref: "#/components/schemas/OfficialDeclaration"
            votes:
              type: array
              items:
                $ref: "#/components/schemas/OfficialVote"
            organi:
              type: array
              items:
                $ref: "#/components/schemas/OfficialOrgano"
    DeliberaVoteEntry:
      type: object
      required: [officialId, name, slug, role, vote]
      properties:
        officialId:
          type: integer
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        group:
          type: ["string", "null"]
        vote:
          type: string
          enum: [favorevole, contrario, astenuto, assente]
    DeliberaVotes:
      type: object
      required: [delibera, tally, votes]
      properties:
        delibera:
          $ref: "#/components/schemas/Publication"
        tally:
          type: object
          required: [favorevole, contrario, astenuto, assente]
          properties:
            favorevole:
              type: integer
            contrario:
              type: integer
            astenuto:
              type: integer
            assente:
              type: integer
        votes:
          type: array
          items:
            $ref: "#/components/schemas/DeliberaVoteEntry"
    OfficialInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        roleTitle:
          type: string
        group:
          type: string
        status:
          type: string
        appointmentDate:
          type: string
        biography:
          type: string
        activities:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
                minLength: 1
              description:
                type: string
              date:
                type: string
        remunerations:
          type: array
          items:
            type: object
            required: [year, type]
            properties:
              year:
                type: integer
              amount:
                type: number
              type:
                type: string
                minLength: 1
              note:
                type: string
        declarations:
          type: array
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
                minLength: 1
              date:
                type: string
              content:
                type: string
              url:
                type: string
        votes:
          type: array
          items:
            type: object
            required: [publicationId, vote]
            properties:
              publicationId:
                type: integer
              vote:
                type: string
                enum: [favorevole, contrario, astenuto, assente]
...